refactor: update DeleteNurseryDialog with unified error handling
- Apply consistent error handling using new error class hierarchy
- Update test cases to match new error structure
- Maintain component functionality while improving error consistency
